+++ NEWS.md
@@ -1,9 +1,17 @@
 # FSA 0.6.20 ongoing
-* `gompFuns()`: Modified.  Added `type="Troynikov1"` and  `type="Troynikov1"`.
-* `logisticFuns()`: Modified.  Changed `type="Richards"` to `type="Karkach"`.
-* `logisticModels()`: Modified.  Added "Karkach" model.
-* `schnute()`: Modified.  Fixed bugs with the way `t1=` and `t3=` are handled.
+* `gompFuns()`: Deleted.
+* `gompModels()`: Deleted.
+* `GompertzFuns()`: Added.  Replaced `gompFuns()`.  Added `type="Troynikov1"` and  `type="Troynikov1"`.
+* `GompertzModels()`: Added.  Replaced `gompModels()`.  Added `cex=` and `type=`.
+* `logisticFuns()`: Modified.  Changed `type="Richards"` to `type="Karkach"`.  Added `type=HaddonI`.
+* `logisticModels()`: Modified.  Added "Karkach" model.  Added `cex=` and `type=`.
+* `RichardsFuns()`: Modified.  Added two more parameterizations from Tjorve and Tjorve (2010).
+* `schnute()`: Deleted.
+* `schnuteModels()`: Deleteed.  
+* `Schnute()`: Added.  Replaced `schnute()`.  Fixed bugs with the way `t1=` and `t3=` are handled.
+* `SchnuteModels()`: Added.  Replaced `schnuteModels()`.  Added `cex=`.
 * `vbFuns()`: Modified.  Added `type="Polacheck"` which is equivalent to `type="Laslett"`.  Added a new reference in the help file.
+* `vbModels()`: Modified.  Added `cex=` and `type=`.
 
 # FSA 0.6.19 Jun15
 * `RichardsFuns()`: Added.